+++ PatternAssociator/UserControls/PatternGrid.xaml.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
+using System.Windows.Controls.Primitives;
 using System.Windows.Data;
 using System.Windows.Documents;
 using System.Windows.Input;
@@ -23,6 +24,16 @@ namespace PatternAssociator.UserControls
         public PatternGrid()
         {
             InitializeComponent();
+            InitializeGrid();
+        }
+
+        //initializes every cell of Pattern (UniformGrid) with a new ToggleButton
+        public void InitializeGrid()
+        {
+            Enumerable
+                .Range(0, Pattern.Rows * Pattern.Columns)
+                .ToList()
+                .ForEach(index => Pattern.Children.Add(new ToggleButton() { Content = "♥" }));
         }
     }
 }