+++ Changelog.md
@@ -8,6 +8,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Well, everything
 
+## [0.1.0] - 2018-05-19
+### Added 
+
+* Pages to train, query and reproduce patterns
+* UserControl to reuse a UniformGrid of ToggleButtons, visualizing patterns represented by binary vectors
+* Horrible Extension to convert a UIElementCollection containing only ToggleButtons into a binary vector
+* Simplified AutoAssociator, maybe, some day, this can be replaced with BrainMaker
+
+### Removed
+
+* Reference to BrainMaker
+
 ## [0.0.1] - 2018-05-17
 ### Added